refactor: split details section into two separated ones
