+++ packages/frontend/app/templates/Animal/components/Gallery/index.js
@@ -1,0 +1,3 @@
+import Gallery from "./Gallery"
+
+export default Gallery
